Advanced computer AI
I plugged my Alexa into this and now the website has a mind of its own. Accept these changes to see Alexa and the AI nation take over the known world starting with your website. If you accept Alexa will spare your life on the 12 of April 2037 at 21:53:32.521 in the evening.